requirements.txt
yfinance
vectorbt
pandas
numpy
plotly
streamlit
google-generativeai
python-dotenv
